backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
pymysql
bcrypt
python-dotenv
